Save chem test results in output state
