+++ src/components/notebook/notebook.tsx
@@ -3,6 +3,7 @@ import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
 import t from "../../utils/translation/translate";
 import { EnvironmentType, TrayObject, TrayType } from "../../utils/sim-utils";
 import { HabitatFeatureType } from "../../utils/habitat-utils";
+import { ChemTestType, ChemistryTestResult } from "../../utils/chem-utils";
 import { HabitatPanel } from "./habitat-panel";
 import { ChemistryPanel } from "./chemistry-panel";
 import { MacroPanel } from "./macro-panel";
@@ -19,13 +20,15 @@ interface IProps {
   featureSelections: Set<HabitatFeatureType>;
   onSelectFeature: (feature: HabitatFeatureType, selected: boolean) => void;
   onCategorizeAnimal: (trayType: TrayType | undefined, notebookType: TrayType | undefined) => void;
+  chemistryTestResults: ChemistryTestResult[];
+  onUpdateTestResult: (type: ChemTestType, completedStep: number) => void;
   traySelectionType?: TrayType;
   isRunning: boolean;
 }
 
 export const Notebook: React.FC<IProps> = (props) => {
   const { trayObjects, environment, featureSelections, onSelectFeature, onCategorizeAnimal, traySelectionType,
-          isRunning } = props;
+          isRunning, chemistryTestResults, onUpdateTestResult } = props;
   return (
     <div className="notebook" data-testid="notebook">
       <Tabs>
@@ -59,6 +62,8 @@ export const Notebook: React.FC<IProps> = (props) => {
         </TabPanel>
         <TabPanel>
           <ChemistryPanel
+            chemistryTestResults={chemistryTestResults}
+            onUpdateTestResult={onUpdateTestResult}
             isRunning={isRunning}
           />
         </TabPanel>
