+++ src/utils/chem-utils.ts
